chabot_2/requirements.txt
python-dotenv

langchain-community
langchain-core
langchain-groq
langchain-huggingface

faiss-cpu

sentence-transformers
huggingface-hub
transformers
torch

numpy
